backend/py/requirements.txt
Flask
keras
Keras_Preprocessing
numpy
Pillow
pymongo
python-dotenv
scipy
tensorflow==2.12.0
tensorflow_hub==0.12.0